refactor: change to advanced parser
