+++ aws_lambda_powertools/utilities/advanced_parser/envelopes/base.py
