+++ aws_lambda_powertools/utilities/advanced_parser/schemas/sqs.py
