+++ aws_lambda_powertools/utilities/advanced_parser/validator.py
@@ -4,7 +4,7 @@ from typing import Any, Callable, Dict
 from pydantic import BaseModel, ValidationError
 
 from aws_lambda_powertools.middleware_factory import lambda_handler_decorator
-from aws_lambda_powertools.utilities.validation.envelopes.base import BaseEnvelope
+from aws_lambda_powertools.utilities.advanced_parser.envelopes.base import BaseEnvelope
 
 logger = logging.getLogger(__name__)
